+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [core] Added endpoint to remove Docker network
 - [core] On startup, core now creates and attaches to its own Docker network, and then disconnects from all other networks
 - [core] Add curl to dev container image
+- [core] Carve out exception to not restart API during tests
 
 ## Changed